Refactor game's dir structure
